requirements.txt
django-environ==0.12.0
django==5.2
djangorestframework==3.16
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
django-celery-beat==2.8.1
django-celery-results==2.6.0
redis==6.1.0
gunicorn==23.0.0